clothes-service/requirements.txt
Django==5.2.11
djangorestframework==3.16.0
